move `MetlogClientHolder` class (and `CLIENT_HOLDER` instance) to its own module.
